+++ test/test_project_utils.py
@@ -3,9 +3,8 @@ import io
 import os
 import pytest
 
-from dcicutils.env_utils import EnvUtils
 from dcicutils.misc_utils import ignorable, override_environ, local_attrs, StorageCell
-from dcicutils.qa_utils import MockFileSystem
+from dcicutils.qa_utils import MockFileSystem, printed_output
 from dcicutils.project_utils import (
     ProjectNames, C4ProjectNames, Project, C4Project, ProjectRegistry, C4ProjectRegistry,
 )
@@ -13,21 +12,30 @@ from unittest import mock
 from dcicutils import project_utils as project_utils_module
 
 
+def app_project_cell_value():
+    return project_utils_module._SHARED_APP_PROJECT_CELL.value  # noQA - testing access to protected member
+
+
+def set_app_project_cell_value(value):
+    project_utils_module._SHARED_APP_PROJECT_CELL.value = value  # noQA - testing access to protected member
+
+
 @contextlib.contextmanager
 def project_registry_test_context(registry=True):
+    temp_cell = StorageCell(initial_value=None)
     attrs = {'APPLICATION_PROJECT_HOME': None, 'PYPROJECT_TOML_FILE': None, 'PYPROJECT_TOML': None,
-             'POETRY_DATA': None, '_shared_app_project_cell': StorageCell(initial_value=None),
-             '_PYPROJECT_NAME': None}
+             'POETRY_DATA': None, '_PYPROJECT_NAME': None}
     if registry:
         attrs['REGISTERED_PROJECTS'] = {}
     with local_attrs(ProjectRegistry, **attrs):
-        yield
+        with mock.patch.object(project_utils_module, '_SHARED_APP_PROJECT_CELL', temp_cell):
+            yield
 
 
 def test_project_registry_register():
 
     mfs = MockFileSystem()
-    with mfs.mock_exists_open_remove():
+    with mfs.mock_exists_open_remove_abspath_getcwd_chdir():
         with project_registry_test_context():
 
             @ProjectRegistry.register('foo')
@@ -104,7 +112,7 @@ def test_project_registry_register_snovault_scenario():
     # These values don't have to be precisely what snovault uses, just plausibly things it might
 
     mfs = MockFileSystem()
-    with mfs.mock_exists_open_remove():
+    with mfs.mock_exists_open_remove_abspath_getcwd_chdir():
         with project_registry_test_context():
 
             # Initialization code expects this filename to be precise
@@ -165,7 +173,7 @@ def test_project_registry_register_cgap_scenario():
     # These values don't have to be precisely what snovault and cgap-portal use, just plausibly things they might
 
     mfs = MockFileSystem()
-    with mfs.mock_exists_open_remove():
+    with mfs.mock_exists_open_remove_abspath_getcwd_chdir():
         with project_registry_test_context():
 
             # Initialization code expects this filename to be precise
@@ -328,7 +336,7 @@ def test_project_registry_make_project():
                 ProjectRegistry._make_project()
             assert str(exc.value) == "ProjectRegistry.PROJECT_NAME not initialized properly."
 
-            ProjectRegistry.initialize_pyproject_name(pyproject_name='foo')
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foo')
             assert isinstance(ProjectRegistry._make_project(), FooProject)
 
             with pytest.raises(Exception) as exc:
@@ -336,7 +344,7 @@ def test_project_registry_make_project():
             assert str(exc.value) == "Registered pyproject 'foo' (FooProject) is not a subclass of C4Project."
 
             ProjectRegistry._PYPROJECT_NAME = 'bogus'  # Mock up a bad initialization
-            ProjectRegistry._shared_app_project_cell.value = None  # Clear cache
+            set_app_project_cell_value(None)  # Clear cache
             with pytest.raises(Exception) as exc:
                 ProjectRegistry._make_project()
             assert str(exc.value) == "Missing project class for pyproject 'bogus'."
@@ -360,28 +368,9 @@ def test_declare_project_registry_class_wrongly():
         Project._PROJECT_REGISTRY_CLASS = old_value
 
 
-def test_app_project_bad_initialization():
-
-    mfs = MockFileSystem()
-    with mfs.mock_exists_open_remove():
-        with project_registry_test_context():
-
-            @ProjectRegistry.register('foo')
-            class FooProject(Project):
-                NAMES = {"NAME": "foo"}
-
-            ProjectRegistry.initialize_pyproject_name(pyproject_name='foo')
-            app_project = FooProject.app_project_maker()
-            project = app_project()
-            project._names = None  # simulate screwing up of initialization
-            with pytest.raises(Exception) as exc:
-                print(project.names)
-            assert str(exc.value) == "<FooProject> failed to initialize correctly."
-
-
 def test_project_registry_class():
 
-    old_registry_class = Project.PROJECT_REGISTRY_CLASS
+    old_registry_class = Project._PROJECT_REGISTRY_CLASS
 
     try:
 
@@ -404,7 +393,7 @@ def test_project_registry_class():
     finally:
 
         # We were supposed to get an error before any side effect happened, but we'll be careful just in case.
-        Project.PROJECT_REGISTRY_CLASS = old_registry_class
+        Project._PROJECT_REGISTRY_CLASS = old_registry_class
 
 
 def test_project_filename():
@@ -417,7 +406,7 @@ def test_project_filename():
             class FooProject(Project):
                 NAMES = {"NAME": "foo"}
 
-            ProjectRegistry.initialize_pyproject_name(pyproject_name='foo')
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foo')
             app_project = FooProject.app_project_maker()
             project = app_project()
 
@@ -429,7 +418,7 @@ def test_project_filename():
                 mock_resource_filename.side_effect = mocked_resource_filename
                 assert project.project_filename('xyz') == "<foo-home>/xyz"
 
-            ProjectRegistry._shared_app_project_cell.value = None  # Mock failure to initialize
+            set_app_project_cell_value(None)  # Mock failure to initialize
 
             with pytest.raises(Exception) as exc:
                 print(project.project_filename("foo"))
@@ -487,6 +476,29 @@ def test_bad_pyproject_names():
                                       " The class can still be SMaHTProject.")
 
 
+def test_project_initialize():
+
+    mfs = MockFileSystem()
+    with mfs.mock_exists_open_remove():
+        with project_registry_test_context():
+
+            @ProjectRegistry.register('foo')
+            class FooProject(Project):
+                NAMES = {"NAME": "foo"}
+
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foo')
+            project = FooProject.initialize()
+            assert isinstance(project, FooProject)
+
+            assert FooProject.initialize() == project  # Doing it again gets the same one
+
+            new_project = Project.initialize(force=True)  # Creates a new instance of the app_project()
+
+            assert isinstance(new_project, FooProject)
+            assert new_project is not project
+            assert new_project.NAME == project.NAME
+
+
 def test_project_registry_initialize():
 
     mfs = MockFileSystem()
@@ -497,12 +509,11 @@ def test_project_registry_initialize():
             class FooProject(Project):
                 NAMES = {"NAME": "foo"}
 
-            ProjectRegistry.initialize_pyproject_name(pyproject_name='foo')
-            app_project = FooProject.app_project_maker()
-            project = app_project()
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foo')
+            project = ProjectRegistry.initialize()
+            assert isinstance(project, FooProject)
 
-            assert ProjectRegistry.initialize() == project
-            assert ProjectRegistry.initialize() == project
+            assert ProjectRegistry.initialize() == project  # Doing it again gets the same one
 
             new_project = ProjectRegistry.initialize(force=True)  # Creates a new instance of the app_project()
 
@@ -523,13 +534,13 @@ def test_initialize_pyproject_name_ambiguity():
 
             with pytest.raises(Exception) as exc:
                 with override_environ(APPLICATION_PYPROJECT_NAME='alpha'):
-                    ProjectRegistry.initialize_pyproject_name(pyproject_toml_file='pyproject.toml',
-                                                              poetry_data={'name': 'omega',
-                                                                           'packages': [{'include': 'omega'}]})
+                    ProjectRegistry._initialize_pyproject_name(pyproject_toml_file='pyproject.toml',
+                                                               poetry_data={'name': 'omega',
+                                                                            'packages': [{'include': 'omega'}]})
             assert str(exc.value) == "APPLICATION_PYPROJECT_NAME='alpha', but pyproject.toml says it should be 'omega'"
 
 
-def test_project_initialize_app_project():
+def test_app_project_via_registry_method():
 
     mfs = MockFileSystem()
     with mfs.mock_exists_open_remove():
@@ -539,38 +550,244 @@ def test_project_initialize_app_project():
             class SuperProject(C4Project):
                 NAMES = {"NAME": "super"}
 
-            ProjectRegistry.initialize_pyproject_name(pyproject_name='super')
-            project = SuperProject.initialize_app_project()
+            with printed_output() as printed:
+                app_project = C4ProjectRegistry.app_project_maker()
+                assert printed.lines == []
 
-            with mock.patch.object(EnvUtils, "init") as mock_init:
-                proj1 = project.initialize_app_project()
-                assert isinstance(proj1, SuperProject)
-                mock_init.assert_called_with()
+            C4ProjectRegistry._initialize_pyproject_name(pyproject_name='super')
 
-                mock_init.reset_mock()
+            proj1 = app_project()
+            assert app_project_cell_value() is not None
+            proj2 = app_project()
+            assert proj1 is proj2
 
-                ProjectRegistry._shared_app_project_cell.value = None   # decache
-                proj2 = project.initialize_app_project(initialize_env_utils=False)
-                assert isinstance(proj2, SuperProject)
-                assert proj1 is not proj2
-                mock_init.assert_not_called()
 
-                mock_init.reset_mock()
+def test_app_project_via_project_method():
 
-                ProjectRegistry._shared_app_project_cell.value = None   # decache
-                # Essentially the same as project.initialize_app_project
-                app_project = project.app_project_maker()
-                app_project(initialize=True, initialization_options={'initialize_env_utils': False})
-                assert isinstance(proj2, SuperProject)
-                assert proj1 is not proj2
-                mock_init.assert_not_called()
+    mfs = MockFileSystem()
+    with mfs.mock_exists_open_remove():
+        with project_registry_test_context():
 
-                mock_init.reset_mock()
+            @C4ProjectRegistry.register('super')
+            class SuperProject(C4Project):
+                NAMES = {"NAME": "super"}
 
-                ProjectRegistry._shared_app_project_cell.value = None   # decache
-                # Essentially the same as project.initialize_app_project
-                app_project = project.app_project_maker()
-                app_project(initialize=True, initialization_options={'initialize_env_utils': True})
-                assert isinstance(proj2, SuperProject)
-                assert proj1 is not proj2
-                mock_init.assert_called_with()
+            print(f"SuperProject.PROJECT_REGISTRY_CLASS={SuperProject.PROJECT_REGISTRY_CLASS.__name__}")
+
+            with printed_output() as printed:
+                app_project = SuperProject.app_project_maker()
+                assert printed.lines == [
+                    ('SuperProject.app_project_maker() called. This class method has been deprecated.'
+                     ' Please use C4ProjectRegistry.app_project_maker() instead.')
+                ]
+
+            C4ProjectRegistry._initialize_pyproject_name(pyproject_name='super')
+
+            proj1 = app_project()
+            assert app_project_cell_value() is not None
+            proj2 = app_project()
+            assert proj1 is proj2
+
+
+def test_app_project_bad_initialization():
+
+    mfs = MockFileSystem()
+    with mfs.mock_exists_open_remove():
+        with project_registry_test_context():
+
+            @ProjectRegistry.register('foo')
+            class FooProject(Project):
+                NAMES = {"NAME": "foo"}
+
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foo')
+            app_project = FooProject.app_project_maker()
+            project = app_project()
+            project._names = None  # simulate screwing up of initialization
+            with pytest.raises(Exception) as exc:
+                print(project.names)
+            assert str(exc.value) == "<FooProject> failed to initialize correctly."
+
+
+MISSING = '<missing>'
+
+
+@pytest.mark.parametrize("verbose", [MISSING, None, True, False])
+@pytest.mark.parametrize("detailed", [MISSING, None, True, False])
+def test_project_registry_show_herald(verbose, detailed):
+
+    print(f"Testing verbose={verbose}, detailed={detailed}")
+
+    mfs = MockFileSystem()
+    with mfs.mock_exists_open_remove_abspath_getcwd_chdir():
+        with project_registry_test_context():
+
+            @ProjectRegistry.register('foo')
+            class FooProject(Project):
+                NAMES = {"NAME": "foo"}
+            ignorable(FooProject)
+
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foo')
+
+            with printed_output() as printed:
+
+                options = {}
+                if detailed is not MISSING:
+                    options['detailed'] = detailed
+                if verbose is not MISSING:
+                    options['verbose'] = verbose
+                print(f" Initialization options: {options}")
+                ProjectRegistry.initialize(**options)
+
+                if verbose is MISSING or verbose is None:
+                    verbose = ProjectRegistry.PROJECT_INITIALIZE_VERBOSE
+
+                if detailed is MISSING or detailed is None:
+                    detailed = ProjectRegistry.PROJECT_INITIALIZE_DETAILED
+
+                print(f" Expected behavior as if verbose={verbose}, detailed={detailed}")
+
+                if not verbose:
+                    assert printed.lines == []
+                elif detailed:
+                    assert printed.lines == [
+                        "",
+                        "==========================================================================================",
+                        "APPLICATION_PROJECT_HOME == '/home/mock'",
+                        "PYPROJECT_TOML_FILE == None",
+                        "PYPROJECT_NAME == 'foo'",
+                        "Project.app_project == ProjectRegistry.app_project == app_project() == <FooProject>",
+                        "app_project().APP_NAME == 'foo'",
+                        "app_project().APP_PRETTY_NAME == 'Foo'",
+                        "app_project().NAME == 'foo'",
+                        "app_project().PACKAGE_NAME == 'foo'",
+                        "app_project().PRETTY_NAME == 'Foo'",
+                        'app_project().PYPI_NAME == None',
+                        "app_project().PYPROJECT_NAME == 'foo'",
+                        "app_project().REPO_NAME == 'foo'",
+                        "==========================================================================================",
+                    ]
+                else:
+                    assert printed.lines == ["FooProject initialized with name 'foo'."]
+
+
+@pytest.mark.parametrize("options_to_test,expected_caveats", [
+    ({},
+     "with name 'foobar'"),
+    ({"PRETTY_NAME": "FooBar"},  # FooBar differs only in case, so won't be flagged
+     "with name 'foobar'"),
+    ({"PRETTY_NAME": "Foo+Bar"},  # Foo+Bar differs only in special characters, so won't be flagged
+     "with name 'foobar'"),
+    ({"PRETTY_NAME": "Fubar"},  # 'Fubar' uses a different spelling so will be noted as different
+     "with name 'foobar' and notable alias 'Fubar'"),
+    ({"PRETTY_NAME": "Fubar", "REPO_NAME": "foo_bar"},  # 'Fubar' uses different spelling but 'foo_bar' does not.
+     "with name 'foobar' and notable alias 'Fubar'"),
+    ({"PRETTY_NAME": "Super Foobar", "REPO_NAME": "FuBar"},  # Both 'FuBar' and 'Super Foobar' are spelled differently.
+     "with name 'foobar' and notable aliases 'FuBar' and 'Super Foobar'"),
+])
+def test_project_registry_show_herald_detailed(options_to_test, expected_caveats):
+
+    mfs = MockFileSystem()
+    with mfs.mock_exists_open_remove_abspath_getcwd_chdir():
+        with project_registry_test_context():
+
+            @ProjectRegistry.register('foobar')
+            class FoobarProject(Project):
+                NAMES = {"NAME": "foobar", **options_to_test}
+            ignorable(FoobarProject)
+
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foobar')
+
+            with printed_output() as printed:
+
+                ProjectRegistry.initialize(verbose=True, detailed=False)
+
+                assert printed.lines == [f"FoobarProject initialized {expected_caveats}."]
+
+
+def test_project_registry_show_herald_consistency():
+
+    mfs = MockFileSystem()
+    with mfs.mock_exists_open_remove_abspath_getcwd_chdir():
+        with project_registry_test_context():
+
+            @ProjectRegistry.register('foobar')
+            class FoobarProject(Project):
+                NAMES = {"NAME": "foobar"}
+            ignorable(FoobarProject)
+
+            ProjectRegistry._initialize_pyproject_name(pyproject_name='foobar')
+
+            rogue_project = FoobarProject()
+
+            # We have to work pretty hard to mess up the consistency, but this should do it.
+            # The value of app_project as an instance property here overrides the class property
+            # that is supposed implement sharing between all the clasess.
+            rogue_project.app_project = rogue_project
+
+            with pytest.raises(Exception) as exc:
+
+                rogue_project.show_herald()
+
+            assert str(exc.value) == "Project consistency check failed."
+
+
+def test_project_names_items():
+
+    names_object = ProjectNames(PYPROJECT_NAME='foo', NAME='foo')
+    items = names_object.items()
+    actual_keys = []
+    for k, v in items:
+        assert isinstance(k, str)
+        assert isinstance(v, str) or (k == 'PYPI_NAME' and v is None)
+        actual_keys.append(k)
+    expected_keys = ['APP_NAME', 'APP_PRETTY_NAME', 'NAME', 'PACKAGE_NAME',
+                     'PRETTY_NAME', 'PYPI_NAME', 'PYPROJECT_NAME', 'REPO_NAME']
+    assert actual_keys == expected_keys
+
+
+def test_find_notable_aliases():
+
+    names = ProjectNames(PYPROJECT_NAME='foo', NAME='foo')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = []
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = ProjectNames(PYPROJECT_NAME='foobar', NAME='foobar', PRETTY_NAME='FuBar')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['FuBar']
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = ProjectNames(PYPROJECT_NAME='foobar', NAME='foobar', PRETTY_NAME='FuBar', PYPI_NAME='Foo Bar')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['FuBar']
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = ProjectNames(PYPROJECT_NAME='encoded', NAME='cgap-portal')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['encoded']
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = C4ProjectNames(PYPROJECT_NAME='encoded', NAME='cgap-portal')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['cgap', 'encoded']
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = ProjectNames(PYPROJECT_NAME='dcicsnovault', NAME='snovault')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['dcicsnovault']
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = C4ProjectNames(PYPROJECT_NAME='dcicsnovault', NAME='snovault')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['dcicsnovault']
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = ProjectNames(PYPROJECT_NAME='encoded-core', NAME='encoded-core')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = []
+    assert actual_notable_aliases == expected_notable_aliases
+
+    names = C4ProjectNames(PYPROJECT_NAME='encoded-core', NAME='encoded-core')
+    actual_notable_aliases = names.find_notable_aliases()
+    expected_notable_aliases = ['core']
+    assert actual_notable_aliases == expected_notable_aliases