+++ CHANGELOG.rst
@@ -10,11 +10,25 @@ Change Log
 7.5.0
 =====
 
+* In ``lang_utils``:
+
+  * Teach ``EnglishUtils.string_pluralize`` about words ending in ``-ses`` because ``cgap-portal`` needs this.
+
 * New module ``project_utils`` with support for Project mechanism.
 
   * New decorators ``ProjectRegistry`` and ``C4ProjectRegistry``
 
   * New class ``Project`` and ``C4Project``
+
+* In ``qa_utils``:
+
+  * In class ``MockFileSystem``:
+
+    * New method ``abspath``
+    * New method ``chdir``
+    * New method ``expanduser``
+    * New method ``getcwd``
+    * New method ``mock_exists_open_remove_abspath_getcwd_chdir`` (context manager)
 
 
 7.4.3